Modify Assertions, DriverFactory Architecture

+++ src/main/java/driverfactory/webdriver/WebDriver.java
@@ -1,28 +1,36 @@
-package driverfactory;
+package driverfactory.webdriver;
 
 import assertions.Assertions;
 import browseractions.BrowserActions;
+import com.browserstack.local.Local;
 import constants.DriverType;
 import constants.EnvType;
-import driverfactory.localdriver.*;
+import driverfactory.webdriver.localdriver.DriverFactory;
 import elementactions.ElementActions;
 import org.openqa.selenium.remote.DesiredCapabilities;
 import org.openqa.selenium.remote.RemoteWebDriver;
 import org.openqa.selenium.support.ThreadGuard;
 import org.testng.Reporter;
+import utilities.JSONFileHandler;
 import utilities.LoggingManager;
+
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.util.HashMap;
+
 import static tools.properties.PropertiesHandler.*;
 
-public class WebDriver{
+public class WebDriver {
 
     private final ThreadLocal<org.openqa.selenium.WebDriver> driverThreadLocal = new ThreadLocal<>();
+    private final Local bsLocal = new Local();
+
+    JSONFileHandler json = new JSONFileHandler("parallel.conf.json");
 
     public WebDriver() {
         LoggingManager.info("Initializing WebDriver.....");
         createWebDriver();
-        if(driverThreadLocal.get() == null){
+        if (driverThreadLocal.get() == null) {
             createWebDriver();
         }
     }
@@ -35,11 +43,16 @@ public class WebDriver{
         if (EnvType.valueOf(getPlatform().environmentType()) == EnvType.GRID) {
             gridInit();
         }
+
+        if (EnvType.valueOf(getPlatform().environmentType()) == EnvType.CLOUD) {
+            cloudInit();
+        }
+
         LoggingManager.info("CURRENT THREAD: " + Thread.currentThread().getId() + ", " + "DRIVER = " + getDriver());
     }
 
 
-    public void localDriverInit() {
+    private void localDriverInit() {
         String baseURL = getCapabilities().baseURL();
         String browserName = Reporter.getCurrentTestResult().getTestClass().getXmlTest().getParameter("browserName");
         LoggingManager.info("Starting " + browserName + " Driver Locally in " + getCapabilities().executionMethod() + " mode");
@@ -53,7 +66,7 @@ public class WebDriver{
 
     }
 
-    public void gridInit() {
+    private void gridInit() {
         String baseURL = getCapabilities().baseURL();
         DesiredCapabilities capabilities = new DesiredCapabilities();
         String browserName = Reporter.getCurrentTestResult().getTestClass().getXmlTest().getParameter("browserName");
@@ -71,6 +84,34 @@ public class WebDriver{
         }
     }
 
+    private void cloudInit() {
+        // You can also set an environment variable - "BROWSERSTACK_ACCESS_KEY".
+        HashMap<String, String> bsLocalArgs = new HashMap<>();
+
+        bsLocalArgs.put("user",json.getData("user"));
+        bsLocalArgs.put("key", json.getData("key"));
+        bsLocalArgs.put("server", json.getData("server"));
+        bsLocalArgs.put("browserName", "chrome");
+        bsLocalArgs.put("os", "Windows");
+        bsLocalArgs.put("osVersion", "10");
+        bsLocalArgs.put("browserstack.local", "true");
+        // Starts the Local instance with the required arguments.
+        try {
+            bsLocal.start(bsLocalArgs);
+            LoggingManager.info("Start Running on BrowserStack Grid......");
+        } catch (Exception e) {
+            LoggingManager.error("Failed to Start BrowserStack Instance, " + e.getMessage());
+        }
+        // Checks if BrowserStack local instance is running.
+        try {
+            LoggingManager.info(bsLocal.isRunning());
+        } catch (Exception e) {
+            LoggingManager.error("BrowserStack Connection isn't started, " + e.getMessage());
+        }
+        // Your test code goes here, from creating the driver instance till the end, i.e. driver.quit.
+        // Stops the Local instance.
+    }
+
 
     private void setDriver(org.openqa.selenium.WebDriver driver) {
         driverThreadLocal.set(driver);
@@ -81,7 +122,7 @@ public class WebDriver{
     }
 
     public org.openqa.selenium.WebDriver getDriver() {
-        if(driverThreadLocal.get() == null){
+        if (driverThreadLocal.get() == null) {
             createWebDriver();
         }
         assert driverThreadLocal.get() != null;
@@ -90,21 +131,30 @@ public class WebDriver{
 
     public void quit() {
         LoggingManager.info("Quitting Driver.....");
-        assert driverThreadLocal.get() != null;
-        driverThreadLocal.get().manage().deleteAllCookies();
-        driverThreadLocal.get().quit();
-        driverThreadLocal.remove();
+        if (EnvType.valueOf(getPlatform().environmentType()) == EnvType.CLOUD){
+            try {
+                bsLocal.stop();
+            } catch (Exception e) {
+                LoggingManager.error("Failed to stop BrowserStack instance, " + e.getMessage());
+            }
+        }
+        else {
+            assert driverThreadLocal.get() != null;
+            driverThreadLocal.get().manage().deleteAllCookies();
+            driverThreadLocal.get().quit();
+            driverThreadLocal.remove();
+        }
     }
 
-    public ElementActions element(){
+    public ElementActions element() {
         return new ElementActions(getDriver());
     }
 
-    public BrowserActions browser(){
+    public BrowserActions browser() {
         return new BrowserActions(getDriver());
     }
 
-    public Assertions assertThat(){
+    public Assertions assertThat() {
         return new Assertions(getDriver());
     }
